Solve background bug on empty password instructions
I eliminated a bug where an empty background square
was shown after clearing the password fields. This
happened due to the recent change where I added a
dark background to the password instructions for
better readability on small screens. I also fixed
a bug where the "check" or "X" mark wasn't properly
removed on the password confirm field when clearing
the password field.