Fix required instructions and small changes
I fixed an issue where the "Fields with an * are
required" paragraph behaving unpredictable. This was
due to the "position" attribute still being set to
"absolute" which was a left-over from an earlier
version. I now re-integrated the paragraph into the
page-flow.
Furthermore I added a drop-shadow effect to the
framed logo to switch from red to green upon filling
out the form correctly.
I also changed the password instructions' font-color
to white for better readability.